Update load up anh

+++ Climate_PE/src/main/java/com/example/demo/service/PatientService.java
@@ -23,7 +23,7 @@ public class PatientService implements UserDetailsService {
 
 	@Autowired
 	private PatientRepository repository;
-	
+
 	@Autowired
 	private ImageService imageService;
 
@@ -100,18 +100,18 @@ public class PatientService implements UserDetailsService {
 		p.setGender(patientDTO.getGender());
 		p.setPhone(patientDTO.getPhone());
 		p.setAvatar(patientDTO.getAvatar());
-		String avartar = imageService.uploadImage(fileData);
-		if(!avartar.equals("Cannot upload file"))
-		{
-		if(!patientDTO.getAvatar().isEmpty())
-		{
-		imageService.deleteImage(patientDTO.getAvatar());
-		}
-		p.setAvatar(avartar);
-		}
-		if(avartar.equals("Cannot upload file"))
-		{
-			return "Update no success";
+		if (fileData != null) {
+			String avartar = imageService.uploadImage(fileData);
+
+			if (!avartar.equals("Cannot upload file")) {
+				if (!patientDTO.getAvatar().isEmpty()) {
+					imageService.deleteImage(patientDTO.getAvatar());
+				}
+				p.setAvatar(avartar);
+			}
+			if (avartar.equals("Cannot upload file")) {
+				return "Update no success";
+			}
 		}
 		if (!p.getEmail().equals(patientDTO.getEmail())) {
 			p.setEmail(patientDTO.getEmail());
